ds: refactor types

+++ src/libtrader/ds/account/order.rs
@@ -4,8 +4,8 @@ use crate::ds::account::position::PositionType;
 pub struct Order {
     pub action_type: PositionType,
     pub stock_symbol: String,
-    pub stock_price: i64,
-    pub stock_amount: i64,
-    pub stock_filled: i64,
+    pub stock_price: f64,
+    pub stock_amount: isize,
+    pub stock_filled: isize,
     pub is_filled: bool,
 }
